cache stuff

+++ cache_interface/src/pybind11_bindings.cpp
@@ -8,7 +8,7 @@ PYBIND11_MODULE(pylibCacheSim, m) {
     m.doc() = "Python interface for libCacheSim";
 
     py::class_<CacheWrapper>(m, "Cache")
-        .def(py::init<const std::string&, uint64_t>())
+        .def(py::init<const std::string&, int>())
         .def("get", &CacheWrapper::get)
         .def("insert", &CacheWrapper::insert)
         .def("evict", &CacheWrapper::evict)
